+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/rootcause/impl/EntityUtils.java
@@ -23,28 +23,6 @@ public class EntityUtils {
   // metric:     thirdeye:metric:{dataset}:{name}
   // metric dim: thirdeye:metric:{dataset}:{name}:{dimension}
 
-  public enum EntityType {
-    EVENT("thirdeye:event"),
-    METRIC("thirdeye:metric"),
-    TIMERANGE("thirdeye:timerange"),
-    BASELINE("thirdeye:baseline"),
-    UNKNOWN("");
-
-    private final String prefix;
-
-    EntityType(String prefix) {
-      this.prefix = prefix;
-    }
-
-    public String getPrefix() {
-      return prefix;
-    }
-
-    public String formatUrn(String format, Object... values) {
-      return this.prefix + ":" + String.format(format, values);
-    }
-  }
-
   /**
    * Returns the entity type as encoded in a URN
    *
