+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/rootcause/impl/MetricDimensionPipeline.java
@@ -47,7 +47,7 @@ public class MetricDimensionPipeline implements Pipeline {
 
   @Override
   public PipelineResult run(ExecutionContext context) {
-    Set<Entity> contextEntities = EntityUtils.filterContext(context, EntityUtils.EntityType.METRIC);
+    Set<Entity> contextEntities = EntityUtils.filterContext(context, EntityType.METRIC);
 
     TimeRangeEntity current = EntityUtils.getContextTimeRange(context);
     if(current == null) {
