+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/rootcause/impl/TimeRangeEntity.java
@@ -9,13 +9,13 @@ import com.linkedin.thirdeye.rootcause.Entity;
  */
 public class TimeRangeEntity extends Entity {
   public static TimeRangeEntity fromURN(String urn, double score) {
-    EntityUtils.assertType(urn, EntityUtils.EntityType.TIMERANGE);
+    EntityUtils.assertType(urn, EntityType.TIMERANGE);
     String[] parts = urn.split(":");
     return new TimeRangeEntity(urn, 1.0, Long.parseLong(parts[2]), Long.parseLong(parts[3]));
   }
 
   public static TimeRangeEntity fromRange(long start, long end) {
-    String urn = EntityUtils.EntityType.TIMERANGE.formatUrn("%d:%d", start, end);
+    String urn = EntityType.TIMERANGE.formatUrn("%d:%d", start, end);
     return new TimeRangeEntity(urn, 1.0, start, end);
   }
